Add tolerance to pixel grid check (#4)
It turns out USGS elevation rasters are not all on exactly the same pixel grid.
This causes errors when trying to merge without resampling:

ValueError: Rasters must have the same resolution to avoid resampling

The differences are so tiny that they shouldn't require resampling. Add a
tolerance to the pixel grid check.