Merge USGS and Sentinel-2 rasters on disk (#10)
We currently merge rasters in memory. Trying to fetch rasters for large
areas causes out of memory errors.

Rasterio's merge function has a `dst_path` argument that allows rasters
to be merged and written to disk in chunks. This adds a corresponding
`dst_path` argument to the USGS and Sentinel-2 data fetching functions,
to allow fetching rasters that are too big to fit in memory.