Aviva/fix duplicate component ids (#11)
After finding the primary components (ie. raster pixels) which intersect
with our boundaries of interest, we join the table with one that gives
us information on the parent materials for each component. In some
cases, this results in duplicates, where one raster component has
multiple parent materials. I'm not sure why this is the case, but we do
not want this to happen. I've added a warning when it does and then
dropped the duplicates.